remove ModuleInitializerProcessor
